requirements.txt
python-dotenv
discord
discord.py
psycopg2
google-generativeai
python-dateutil